rec_sys/requirements.txt
# Core ML libraries
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
pyarrow>=12.0.0  # For parquet files
fastparquet>=0.8.0

# Vector search and retrieval
faiss-cpu>=1.7.4
redis>=4.5.0

# API serving
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database connections
pymongo>=4.4.0
motor>=3.2.0  # Async MongoDB driver

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Experiment tracking (optional)
wandb>=0.15.0
mlflow>=2.5.0
